Add base neural network class
