+++ src/models/__init__.py
@@ -1,5 +1,5 @@
 from .parsimonious import Persistence
 from .regression import LinearRegression
-from .linear_network import LinearNetwork
+from .neural_networks.linear_network import LinearNetwork
 
 __all__ = ['Persistence', 'LinearRegression', 'LinearNetwork']
