+++ src/models/neural_networks/base.py
@@ -1,59 +1,33 @@
 import numpy as np
-from pathlib import Path
 import random
+from pathlib import Path
 import math
 
 import torch
-from torch import nn
 from torch.nn import functional as F
 
 import shap
 
-from typing import cast, Any, Dict, List, Optional, Tuple, Union
+from typing import Any, Dict, List, Optional, Tuple, Union
 
-from .base import ModelBase
-from .utils import chunk_array
-from .data import DataLoader, train_val_mask
+from ..base import ModelBase
+from ..utils import chunk_array
+from ..data import DataLoader, train_val_mask
 
 
-class LinearNetwork(ModelBase):
+class NNBase(ModelBase):
 
-    model_name = 'linear_network'
-
-    def __init__(self, layer_sizes: Union[int, List[int]],
-                 dropout: float = 0.25,
+    def __init__(self,
                  data_folder: Path = Path('data'),
                  batch_size: int = 1,
                  pred_months: Optional[List[int]] = None,
                  include_pred_month: bool = True) -> None:
         super().__init__(data_folder, batch_size, pred_months, include_pred_month)
 
-        if type(layer_sizes) is int:
-            layer_sizes = cast(List[int], [layer_sizes])
-
-        # to initialize and save the model
-        self.layer_sizes = layer_sizes
-        self.dropout = dropout
-        self.input_size: Optional[int] = None
-
         # for reproducibility
         torch.manual_seed(42)
 
         self.explainer: Optional[shap.DeepExplainer] = None
-
-    def save_model(self):
-
-        assert self.model is not None, 'Model must be trained before it can be saved!'
-
-        model_dict = {
-            'state_dict': self.model.state_dict(),
-            'layer_sizes': self.layer_sizes,
-            'dropout': self.dropout,
-            'input_size': self.input_size,
-            'include_pred_month': self.include_pred_month
-        }
-
-        torch.save(model_dict, self.model_dir / 'model.pkl')
 
     def explain(self, x: Any) -> np.ndarray:
         assert self.model is not None, 'Model must be trained!'
@@ -69,6 +43,9 @@ class LinearNetwork(ModelBase):
                 x[1] = self._one_hot_months(x[1])
 
         return self.explainer.shap_values(x)
+
+    def _initialize_model(self, x_ref: Tuple[torch.Tensor, ...]) -> torch.nn.Module:
+        raise NotImplementedError
 
     def train(self, num_epochs: int = 1,
               early_stopping: Optional[int] = None,
@@ -99,13 +76,10 @@ class LinearNetwork(ModelBase):
                                           pred_months=self.pred_months, to_tensor=True)
 
         # initialize the model
-        if self.input_size is None:
+        if self.model is None:
             x_ref, _ = next(iter(train_dataloader))
-            self.input_size = x_ref[0].contiguous().view(x_ref[0].shape[0], -1).shape[1]
-        self.model: LinearModel = LinearModel(input_size=self.input_size,
-                                              layer_sizes=self.layer_sizes,
-                                              dropout=self.dropout,
-                                              include_pred_month=self.include_pred_month)
+            model = self._initialize_model(x_ref)
+            self.model = model
 
         optimizer = torch.optim.Adam([pam for pam in self.model.parameters()],
                                      lr=learning_rate)
@@ -199,60 +173,3 @@ class LinearNetwork(ModelBase):
     @staticmethod
     def _one_hot_months(indices: torch.Tensor) -> torch.Tensor:
         return torch.eye(14)[indices.long()][:, 1:-1]
-
-
-class LinearModel(nn.Module):
-
-    def __init__(self, input_size, layer_sizes, dropout, include_pred_month):
-        super().__init__()
-
-        self.include_pred_month = include_pred_month
-        if self.include_pred_month:
-            input_size += 12
-        layer_sizes.insert(0, input_size)
-
-        self.dense_layers = nn.ModuleList([
-            LinearBlock(in_features=layer_sizes[i - 1],
-                        out_features=layer_sizes[i], dropout=dropout) for
-            i in range(1, len(layer_sizes))
-        ])
-
-        self.final_dense = nn.Linear(in_features=layer_sizes[-1], out_features=1)
-
-        self.init_weights()
-
-    def init_weights(self):
-        for dense_layer in self.dense_layers:
-            nn.init.kaiming_uniform_(dense_layer.linear.weight.data)
-
-        nn.init.kaiming_uniform_(self.final_dense.weight.data)
-        # http://cs231n.github.io/neural-networks-2/#init
-        # see: Initializing the biases
-        nn.init.constant_(self.final_dense.bias.data, 0)
-
-    def forward(self, x, pred_month=None):
-        # flatten
-        x = x.contiguous().view(x.shape[0], -1)
-        if self.include_pred_month:
-            x = torch.cat((x, pred_month), dim=-1)
-        for layer in self.dense_layers:
-            x = layer(x)
-
-        return self.final_dense(x)
-
-
-class LinearBlock(nn.Module):
-    """
-    A linear layer followed by batchnorm, a ReLU activation, and dropout
-    """
-
-    def __init__(self, in_features, out_features, dropout=0.25):
-        super().__init__()
-        self.linear = nn.Linear(in_features=in_features, out_features=out_features, bias=False)
-        self.relu = nn.LeakyReLU(negative_slope=0.1, inplace=True)
-        self.batchnorm = nn.BatchNorm1d(num_features=out_features)
-        self.dropout = nn.Dropout(dropout)
-
-    def forward(self, x):
-        x = self.relu(self.batchnorm(self.linear(x)))
-        return self.dropout(x)
