+++ tests/models/test_linear_network.py
@@ -4,7 +4,7 @@ import pytest
 import torch
 
 from src.models import LinearNetwork
-from src.models.linear_network import LinearModel
+from src.models.neural_networks.linear_network import LinearModel
 
 from ..utils import _make_dataset
